1292 - move payment method components to a folder
